[feature-5] Data layer was refactored.
