+++ src/data/i-storage-factory.ts
@@ -3,5 +3,5 @@ import { Entity } from './entities/entity';
 import { IStorage } from './i-storage';
 
 export interface IStorageFactory {
-    make<TEntity extends Entity>(): Observable<IStorage<TEntity>>;
+    make<TEntity extends Entity>(): IStorage<TEntity>;
 }
